opengl: render triangulated lines
Workaround for glLineWidth always == 1f on some systems

+++ data/shader/render.shader
@@ -31,7 +31,7 @@ flat SHARED uint fragFlags;
 flat SHARED uint fragTextureWindow;
 
 #ifdef VERTEX_SHADER
-in ivec2 position;
+in vec2 position;
 in uvec3 color;
 in ivec2 texcoord;
 in uint bitcount;
@@ -41,7 +41,7 @@ in uint flags;
 in uint textureWindow;
 
 void main() {
-    vec2 pos = vec2((float(position.x) - displayAreaPos.x) / displayAreaSize.x, (float(position.y) - displayAreaPos.y) / displayAreaSize.y);
+    vec2 pos = vec2((position.x - displayAreaPos.x) / displayAreaSize.x, (position.y - displayAreaPos.y) / displayAreaSize.y);
     // vec2 pos = vec2(position.x / 1024.f, position.y / 512.f);
     fragColor = vec3(float(color.r) / 255.f, float(color.g) / 255.f, float(color.b) / 255.f);
     fragTexcoord = vec2(texcoord.x, texcoord.y);
